[Update] Performance, Reduction of code amount
