+++ package.json
@@ -10,11 +10,14 @@
   },
   "dependencies": {
     "core-js": "^3.9.1",
+    "fibers": "^5.0.0",
     "nuxt": "^2.15.3"
   },
   "devDependencies": {
     "eslint-config-prettier": "^8.1.0",
     "eslint-plugin-prettier": "^3.3.1",
-    "prettier": "^2.2.1"
+    "prettier": "^2.2.1",
+    "sass": "^1.48.0",
+    "sass-loader": "^10.1.1"
   }
 }
